# apis/videos/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an api to show videos",
        "title": "Finpay Task - Video API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/admin/apikey": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get APIKeys"
                ],
                "summary": "Get all youtube apikeys from list",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Add APIKey"
                ],
                "summary": "Add youtube apikey to the list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key to be added",
                        "name": "apikey",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/video/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video Search"
                ],
                "summary": "Get latest videos by query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query term to search",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    }
}